requirements.txt
# Core ML/Audio
torch==2.8.0
torchaudio==2.8.0
transformers==4.56.0

# Numeric/Data Processing
numpy==1.26.4
pandas==2.3.2
pyarrow==19.0.1
psycopg2-binary==2.9.10

# Web Scraping
requests==2.32.5
beautifulsoup4==4.13.5
lxml==6.0.1
playwright==1.48.0

# Cloud Storage
google-cloud-storage==2.10.0
gcsfs==2025.10.0

# Utilities
tqdm==4.67.1
python-dotenv==1.0.1
